Redirect to root URL instead of dashboard
This attempts to fix a bug introduced in Jira 7.2.2 that caused the Jira
"oops" 404 page to display after logging in with SAML.

For some reason I still don't understand, the `sendRedirect` method does
not actually issue a 302 redirect when redirecting to
`/secure/Dashboard.jspa`.  Redirecting to the root URL works fine.  Jira
then immediately redirects to the user's default starting page.